TODO
TODO, if an app or setting or whateva already has been rendered, then
simply show the containing element.
When qr-input AppEvent comes, trigger 'render-app', 'register', etc.
with the qr code as last argument.

When passing arguments to render-app, the first one should be the
containing element, the second the app id, and then more arguments that
will be passed to the app view.